requirements.txt
os
request
json
bs4
